refactor: use arrow functions and remove useless imports
