+++ src/components/Recent/CustomLocationCellRenderer.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Tooltip } from '../Common/Tooltip';
 import styles from './CustomCellRenderer.module.css';
