+++ src/components/Recent/GraphGridItem.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
 import { img } from '../../assets/home';
 import { openFile } from '../../functions/utility';
 import { CardItem } from '../Common/CardItem';
 
-export function GraphGridItem({ id, Caption, ContextData, Description, DateModified, Thumbnail, setIsDisabled }: GraphItem) {
+export const GraphGridItem = ({ id, Caption, ContextData, Description, DateModified, Thumbnail, setIsDisabled }: GraphItem) => {
     const handleClick = (e:MouseEvent) => {
         // freezes the UI 
         setIsDisabled(true);
