+++ src/components/Samples/SamplesGridItem.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
 import { img } from '../../assets/home';
 import { openFile } from '../../functions/utility';
 import { CardItem } from '../Common/CardItem';
 
-export function SamplesGridItem({ FileName, FilePath }: {FileName: string, FilePath: string}) {
+export const SamplesGridItem = ({ FileName, FilePath }: {FileName: string, FilePath: string}) => {
     const handleClick = (e: MouseEvent) => {
         e.preventDefault();
         openFile(FilePath);
