+++ src/functions/utility.ts
@@ -3,7 +3,7 @@
 * @param {string} path | - the location of the file on the system
 */
 export function openFile(path:string) {
-  if (window.chrome.webview !== undefined) {
+  if (window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.OpenFile(path);
   }
 }
@@ -13,7 +13,7 @@ export function openFile(path:string) {
  * @param {string} guidedTour - the type of guided tour to be started
  */
 export function startGuidedTour(guidedTour:string) {
-  if (window.chrome.webview !== undefined) {
+  if (window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.StartGuidedTour(guidedTour);
   }
 }
@@ -23,19 +23,19 @@ export function startGuidedTour(guidedTour:string) {
  * @param {SidebarCommand} value the type of command based on the UI request
  */
 export function sideBarCommand(value: SidebarCommand) {
-  if (value === 'open-file' && window.chrome.webview !== undefined) {
+  if (value === 'open-file' && window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.OpenWorkspace();
   }
-  if (value === 'open-template' && window.chrome.webview !== undefined) {
+  if (value === 'open-template' && window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.ShowTempalte();
   }
-  if (value === 'open-backup-locations' && window.chrome.webview !== undefined) {
+  if (value === 'open-backup-locations' && window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.ShowBackupFilesInFolder();
   }
-  if (value === 'workspace' && window.chrome.webview !== undefined) {
+  if (value === 'workspace' && window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.NewWorkspace();
   }
-  if (value === 'custom-node' && window.chrome.webview !== undefined) {
+  if (value === 'custom-node' && window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.NewCustomNodeWorkspace();
   }
 }
@@ -44,7 +44,7 @@ export function sideBarCommand(value: SidebarCommand) {
  * A call to a backend function requesting the execution of the ShowSampleFilesInFolder function
  */
 export function showSamplesFilesInFolder() {
-  if (window.chrome.webview !== undefined) {
+  if (window.chrome?.webview !== undefined) {
     window.chrome.webview.hostObjects.scriptObject.ShowSampleFilesInFolder();
   }
 }
@@ -52,8 +52,8 @@ export function showSamplesFilesInFolder() {
 /**
  * A call to a backend function requesting to save the current HomePage settings
  */
-export function saveHomePageSettings(settings) {
-  if (window.chrome.webview !== undefined) {
+export function saveHomePageSettings(settings: any) {
+  if (window.chrome?.webview !== undefined) {
     const settingsJson = JSON.stringify(settings);
     window.chrome.webview.hostObjects.scriptObject.SaveSettings(settingsJson);
   }
