--- types/arrow.d.ts
@@ -1,6 +1,0 @@
-type Direction = 'right'|'left';
-type Arrow = {
-  isOpen: boolean;
-  direction?: Direction;
-  color?: string;
-}
