--- types/author-cell.d.ts
@@ -1,4 +1,0 @@
-type AuthorCell = {
-  value: string;
-  row?: any;
-}
