--- types/card-item.d.ts
@@ -1,7 +1,0 @@
-type CardItem = {
-  imageSrc: string;
-  onClick: (e) => void;
-  tooltipContent: string;
-  titleText: string;
-  subtitleText: string;
-}
