--- types/dropdown.d.ts
@@ -1,11 +1,0 @@
-type option = {
-  label: any;
-  value: string;
-}
-
-type Dropdown = {
-  id: string;
-  options: option[];
-  placeholder: any;
-  onSelectionChange: (value: string) => void
-}
