--- types/graph-item.d.ts
@@ -1,9 +1,0 @@
-type GraphItem = {
-  id: string;
-  Caption: string;
-  ContextData: any;
-  Description: string;
-  DateModified: string;
-  Thumbnail: string;
-  setIsDisabled: (disable: boolean) => void;
-}
