--- types/guide.d.ts
@@ -1,7 +1,0 @@
-type Guide = {
-  id: string;
-  Name: string;
-  Description: string;
-  Type: 'test' | '';
-  Thumbnail?: string;
-}
