+++ types/index.d.ts
@@ -1,11 +1,2 @@
-declare module 'src/localization/localization';
-declare module 'src/components/LayoutContainer';
-declare module 'src/components/SettingsContext';
-declare module 'src/components/Sidebar/Sidebar';
 declare module 'react-split-pane';
-declare module 'src/components/MainContent';
-declare module 'src/components/Recent/PageRecent';
-declare module 'src/components/Samples/PageSamples';
-declare module 'src/components/Learning/PageLearning';
 declare module 'react-table';
-declare module 'src/components/Samples/SamplesGrid';
