--- types/modal-item.d.ts
@@ -1,6 +1,0 @@
-
-type ModalItem = {
-  isOpen?: boolean;
-  onClose?: ()=> void;
-  children: ReactNode;
-}
