--- types/name-cell.d.ts
@@ -1,9 +1,0 @@
-type NameCell = {
-  value: string;
-  row: {
-    original: {
-      Thumbnail?: string;
-      Description: string;
-    }
-  }
-}
