--- types/sidebar.d.ts
@@ -1,4 +1,0 @@
-type Sidebar = {
-  onItemSelect: (item: SidebarItem) => void;
-  selectedSidebarItem: SidebarItem;
-}
