--- types/video-carousel-item.d.ts
@@ -1,6 +1,0 @@
-type VideoCarouselItem = {
-  id: string;
-  title: string;
-  videoId: string;
-  description: string;
-}
